+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 Changelog
 =========
 
+## Unreleased
+### Fixed
+- Exception when 'Authorization Bearer' was the only header value passed (no token)
+
 ## [1.0.0-rc.2] - 2018-03-21
 ### Changed
 - Authorization server name to SelfConsumable as it's targeted and more appropriate